refactor(controller): extract recall.ts (R3)
Move the run-scoped artifact-recall cluster out of controller-coordinator-handler.ts
into a new sibling module controller/recall.ts: 6 RECALL_* constants,
buildRecallBlock, collectApproved, runScopedRecall, rankStatus, isBetterStep,
isBetterMcp, MAX_EXTRACT_WINDOWS, relevantExtract. Handler re-exports
runScopedRecall + relevantExtract for import-path stability (test uses them).
cosine import and KnowledgeEntry type removed from handler; resolveByPrecedence
and R1 helpers (extractPrompt, synthMeta, mapOutcome, recordStepControl) stay.
Handler: 1887 → 1690 lines; recall.ts: 218 lines.
